pyproject.toml
[tool.poetry]
name = "aws-graphql-subscription-example"
description = "GraphQL Subscription over AWS API Gateway"
authors = ["Louis Amon <louis@nuage.studio>"]
license = "Proprietary"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.9"
aws-lambda-powertools = { version = "^2.4.0", extras = ["parser", "tracer"] }
boto3 = "^1.26.25"
strawberry-graphql = { version = "^0.159.0", extras = ["fastapi", "pydantic"] }
mangum = "^0.17.0"

[tool.poetry.group.dev.dependencies]
# uvicorn[standard] is required to run a websocket server
uvicorn = { version = "^0.20.0", extras = ["standard"] }
boto3-stubs = { version = "^1.26.83", extras = ["apigatewaymanagementapi"] }
pylint = "^2.17.2"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.poetry.group.infra.dependencies]
pulumi = "^3.63.0"
pulumi-aws = "^5.36.0"
pulumi-docker = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
preview = true

[tool.isort]
line_length = 88
profile = "black"
skip_gitignore = true
known_first_party = ["api", "infra"]
